Update LWI to PHP 8.1 (really)

+++ lwi/public_html/download_res.php
@@ -1,6 +1,6 @@
 <?php
 
-/* 
+/*
  * Copyright (C) 2021 Marcelo C. Pereira <mcper at unicamp.br>
  *
  * This program is free software: you can redistribute it and/or modify
@@ -19,25 +19,31 @@
 
 require '../defaults.php';
 
-$session_id = preg_replace( "/[^\da-z]/i", "", filter_input( INPUT_COOKIE, session_name( ), FILTER_SANITIZE_STRING ) );
+$cookie_id = filter_input( INPUT_COOKIE, session_name( ), FILTER_SANITIZE_FULL_SPECIAL_CHARS );
+if ( ! is_null( $cookie_id ) && is_string( $cookie_id ) ) {
+    $session_id = preg_replace( "/[^\da-z]/i", "", $cookie_id );
+} else {
+    $session_id = "NOCOOKIE";
+}
+
 $session_short_id = substr( $session_id, -6 );
 
 // check if results file exists
 $filename_res = $output_pref . "run-" . $session_short_id . "_*.csv";
-$filename_res = glob( $filename_res ); 
+$filename_res = glob( $filename_res );
 
-if ( count( $filename_res ) === 1 ) {
+if ( is_countable( $filename_res ) && count( $filename_res ) === 1 ) {
     echo $filename_res[ 0 ];
-} elseif ( count( $filename_res ) > 1 ) {
+} elseif ( is_countable( $filename_res ) && count( $filename_res ) > 1 ) {
     $filename_zip = $output_pref . "run-" . $session_short_id . ".zip";
-    
+
     if ( file_exists( $filename_zip ) ) {
         unlink( $filename_zip );
     }
-    
+
     // adjust Windows executable names
     if ( strtoupper( substr( PHP_OS, 0, 3 ) ) === "WIN" ) {
-        $res = exec( "tar -a -c -C " . $output_pref . " -f " . $filename_zip . " " . implode( " ", array_map( basename, $filename_res ) ) );    
+        $res = exec( "tar -a -c -C " . $output_pref . " -f " . $filename_zip . " " . implode( " ", array_map( basename, $filename_res ) ) );
     } else {
         $res = exec( "zip -j " . $filename_zip . " " . implode( " ", $filename_res ) );
     }
@@ -47,7 +53,7 @@ if ( count( $filename_res ) === 1 ) {
     } else {
         echo "NoFile:";
     }
-    
+
 } else {
     echo "NoFile:";
 }
